Move EventDateComparator into eventseries package
